requirements.txt
# PDF Editor — dépendances Python
# Installer avec : pip install -r requirements.txt

# Interface graphique
PySide6>=6.6.0

# Rendu PDF haute qualité (moteur pdfium/Chromium)
pypdfium2>=5.0.0

# Manipulation PDF avancée
pikepdf>=10.0.0

# Manipulation PDF utilitaire (merge, split, encrypt)
pypdf>=6.0.0

# Extraction texte avec layout
pdfplumber>=0.11.0

# OCR (nécessite Tesseract installé séparément)
pytesseract>=0.3.10
Pillow>=10.0.0

# Signature numérique
pyhanko>=0.20.0
pyhanko-certvalidator>=0.26.0

# Packaging (optionnel, pour générer le .exe)
# pyinstaller>=6.0.0
markdown>=3.5

# Conversion image → PDF (lossless JPEG embedding, PNG, TIFF multi-pages)
img2pdf>=0.5.0
